fix: change js to ts files, change filenames to kebabcase, remove oboslete code, move components folder outside of app folder
