--- app/components/icons/youtube.js
@@ -1,10 +1,0 @@
-export const youtubeMask = {
-    id: "mask-youtube",
-    path: `M524.8,178.7c0,0-3.5-24.6-14.2-35.4c-13.5-14.2-28.7-14.3-35.7-15.1c-49.9-3.6-124.8-3.2-124.8-3.2
-    s-74.9-0.4-124.8,3.2c-7,0.8-22.1,0.9-35.7,15.1c-10.7,10.8-14.2,35.4-14.2,35.4s-3.6,28.8-4.2,57.7v27c0.7,28.8,4.2,57.7,4.2,57.7
-    s3.5,24.6,14.2,35.4c13.5,14.2,31.4,13.7,39.3,15.2c28,2.7,118,3.3,121.1,3.3c0,0,75,0.1,124.8-3.5c7-0.8,22.1-0.9,35.7-15.1
-    c10.7-10.8,14.2-35.4,14.2-35.4s3.6-28.8,3.8-57.7v-27C528.4,207.5,524.8,178.7,524.8,178.7z M313.8,299.9V199.4l95.8,50.2
-    L313.8,299.9z`,
-    link: "https://www.youtube.com/"
-  };
-
