--- app/hooks/useIsMobile.js
@@ -1,14 +1,0 @@
-import { useEffect, useState } from "react";
-
-export default function useIsMobile(breakpoint = 768) {
-    const [isMobile, setIsMobile] = useState(false);
-
-    useEffect(() => {
-        const checkMobile = () => setIsMobile(window.innerWidth < breakpoint);
-
-        checkMobile();
-        window.addEventListener('resize', checkMobile);
-        return () => window.removeEventListener('resize', checkMobile);
-    }, [breakpoint])
-    return isMobile;
-};
