+++ app/layout.tsx
@@ -1,6 +1,6 @@
 import { Geist, Geist_Mono } from "next/font/google";
-import Navbar from "./components/Navbar"
-import Footer from "./components/Footer"
+import Navbar from "@/components/navbar"
+import Footer from "@/components/footer"
 import { barlowCondensed } from "../styles/fonts"
 import "../styles/globals.css";
 import { views } from "../lib/views"
