+++ components/parallax-text.tsx
