+++ lib/contentful/image-variants.ts
