+++ tsconfig.json
@@ -22,6 +22,6 @@
       "@/*": ["./*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "app/page.tsx", "lib/contentful.ts", "lib/contentful/contentfulUtils.js", "app/layout.tsx", "next.config.js", "app/components/LightBox.jsx"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "app/page.tsx", "lib/contentful.ts", "lib/contentful/contentfulUtils.js", "app/layout.tsx", "next.config.js"],
   "exclude": ["node_modules"]
 }
